+++ include/compiler/read_num.h
@@ -4,4 +4,4 @@
 
 // Try to read number
 // If it was successful, run the function
-bool ReadNum(char **program_ptr, void function(long num));
+bool ReadNum(char **program_ptr, void function(int num));
