+++ src/compiler/output.c
@@ -19,23 +19,23 @@ void CompileTo_x86_64(){
     // general
     Output = x86_64_Output;
     // c-stack
-    OutputStackPush       = x86_64_StackPush;
+    OutputStackPush  = x86_64_StackPush;
     // c-ctrl
-    OutputCtrlCall        = x86_64_CtrlCall;
+    OutputCtrlCall   = x86_64_CtrlCall;
     // c-tag
-    OutputTagDef          = x86_64_TagDef;
-    OutputTagJmp          = x86_64_TagJmp;
+    OutputTagDef     = x86_64_TagDef;
+    OutputTagJmp     = x86_64_TagJmp;
 }
 
 void (*Output)(Operation operation) = NULL;
 
 // c-stack
-void (*OutputStackPush)(long num) = NULL;
+void (*OutputStackPush)(int num) = NULL;
 
 // c-ctrl
-void (*OutputCtrlCall)(long num) = NULL;
+void (*OutputCtrlCall)(int num) = NULL;
 
 // c-tag
-void (*OutputTagDef)(long num) = NULL;
+void (*OutputTagDef)(int num) = NULL;
 
-void (*OutputTagJmp)(long num) = NULL;
+void (*OutputTagJmp)(int num) = NULL;
